+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to the code will be documented in this file.
 
+## [0.0.6]
+
+Features
+
+- Added ability to color the activityBar, statusBar, and titleBar
+- Added the property `peacock.affectedSettings` to allow affecting one or more of the following:
+  - activityBar
+  - statusBar
+  - titleBar
+
+Refactor
+
+- Refactored all utility functions to `utils.ts`
+
 ## [0.0.5]
 
 Features